Converted floats back to doubles to match project instructions

+++ src/brute-force-algorithm.cpp
@@ -8,6 +8,6 @@ void BruteForceAlgorithm::runBruteFoceAlgorithm() {
 
 }
 
-float BruteForceAlgorithm::getLowestFoundTourCost() {
+double BruteForceAlgorithm::getLowestFoundTourCost() {
 
 }
